requirements.txt
fastapi
uvicorn[standard]
redis
google-generativeai
requests
python-dotenv
